requirements.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
billiard==3.3.0.23
celery==3.1.25
certifi==2018.4.16
cffi==1.11.5
constantly==15.1.0
cryptography==2.3
cssselect==1.0.3
decorator==4.3.0
Django==2.0.4
django-authtools==1.6.0
django-braces==1.12.0
django-celery==3.2.1
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-storages==1.6.6
django-tagulous==0.13.2
easy-thumbnails==2.5
future==0.15.2
gunicorn==19.8.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
jsonpath-rw==1.4.0
kombu==3.0.37
lxml==4.2.3
parsel==1.5.0
Pillow==3.4.2
pip==18.0
ply==3.11
psycopg2==2.7.5
pyasn1==0.4.3
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
pytz==2018.3
queuelib==1.5.0
Scrapy==1.4.0
scrapy-djangoitem==1.1.1
scrapy-splash==0.7.2
scrapyd==1.2.0
service-identity==17.0.0
setuptools==39.2.0
six==1.11.0
sqlparse==0.2.4
Twisted==18.7.0
w3lib==1.19.0
Werkzeug==0.14.1
wheel==0.31.1
zope.interface==4.5.0
